requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
serverless-wsgi==3.1.0
feedparser==6.0.11
anthropic==0.34.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
